extract kvstore into new crate
