+++ Cargo.toml
@@ -67,6 +67,7 @@ members = [
     "genesis",
     "install",
     "keygen",
+    "kvstore",
     "ledger-tool",
     "logger",
     "metrics",
@@ -89,5 +90,5 @@ members = [
     "upload-perf",
     "vote-signer",
     "wallet",
-]
-exclude = ["programs/bpf/rust/noop"]
+    ]
+    exclude = ["programs/bpf/rust/noop"]
