+++ core/src/lib.rs
@@ -37,8 +37,6 @@ pub mod fetch_stage;
 pub mod fullnode;
 pub mod gen_keys;
 pub mod gossip_service;
-#[cfg(feature = "kvstore")]
-pub mod kvstore;
 pub mod leader_confirmation_service;
 pub mod leader_schedule;
 pub mod leader_schedule_utils;
