+++ kvstore/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "solana-kvstore"
+description = "Embedded Key-Value store for solana"
+version = "0.0.1"
+homepage = "https://solana.com/"
+readme = "../README.md"
+repository = "https://github.com/solana-labs/solana"
+authors = ["Solana Maintainers <maintainers@solana.com>"]
+license = "Apache-2.0"
+edition = "2018"
+
+[dependencies]
+bincode = "1.1.2"
+byteorder = "1.3.1"
+chrono = "0.4.0"
+crc = "1.8.1"
+memmap = "0.7.0"
+rand = "0.6.5"
+serde = "1.0.89"
+serde_derive = "1.0.88"
+
