+++ kvstore/src/error.rs
