+++ kvstore/src/io_utils.rs
